2006-05-24  Matthew Gregan  <kinetik@orcon.net.nz>
	* netxx/osutil.{cxx,h}: Add OS-specific strerror function.  This
	continues to call strerror() on POSIX, and tries using
	FormatMessage() and a built-in error message table on Win32.
	* netxx/{accept,datagram,peer,probe_select,recvfrom}.cxx,
	netxx/{resolve_gethostbyname,serverbase,socket,sockopt}.cxx,
	netxx/{streambase,streamserver}.cxx: s/strerror/Netxx::str_error/g.
About Netxx::str_error: why use an internal table of strings?  Doesn't WSAErrStr() give the proper result?
Not sure why this had to be a comment on the revision as well as a private
mail, to be honest.  Here's the response I sent via mail:

Thanks for taking a look at it.  I wrote the code that way because it was
the only sane way I'm aware of doing things.  What Netxx was doing before
was totally wrong.  FormatMessage is not documented to handle errors
returned from WSAGetLastError, so you can't rely on it returning anything
useful... but as of Win2k, it happens to work--so we try to use it, and if
we don't get anything useful back, we use the internal message table.

This should work fairly well--anybody with a modern version of Windows will
see the "official" (albeit from an undocumented source) error in the native
language of their Windows installation, and other users will at least get
something better than "unknown error", which is what strerror() used to give
us.  If anyone wants to go to the effort of making the internal string table
translatable, I won't stop them. :-)

A good reference about this issue is:
http://tangentsoft.net/wskfaq/newbie.html#errmsg

If there was a better way to handle this, I couldn't find it in the
documentation while I was working on this.  I didn't know WSAErrStr
existed--it's not mentioned anywhere in the WSAGetLastError or FormatMessage
pages in MSDN.

Hmm, looking further, it doesn't appear to exist (or, at least, be
documented) in the Win32 API?  It's not mentioned in MSDN or the MinGW
headers at all.  Is it a standard Win32 API, and can you point me in the
right direction?  I'm happy to use something like that if exists and is
generally available on Windows.

So, in summary, the reason I'm not using WSAErrStr is because I didn't know
it existed at the time I wrote the code, and having looked now that it has
been brought to my attention, it is clear that it's not a standard Win32
function, but a function from a utility library shipped with the book
"Windows Sockets Network Programming".