"Fix" race in some netsync tests.
There's nothing that says that the server post-netsync hooks will finish
before the client exits. Tests that rely on the results of these hooks
can fail if they kill the server before the hook finishes, for example
the netsync_hook_errcodes test fails 2-4% of the time on my system.
Put a "sleep(1)" in the kill-server function to always give it enough time.